requirements.txt
api==0.0.7
Django==4.2.19
djangorestframework==3.13.1
